feat(kobo): Move kobo controllers to a new folder
