+++ src/Controller/Kobo/KoboDownloadController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Kobo;
 
 use App\Entity\Book;
 use App\Entity\KoboDevice;
