+++ src/Controller/Kobo/KoboNextReadController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Kobo;
 
 use App\Kobo\Proxy\KoboStoreProxy;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
